Update README.md
Not content updates.  But interesting insight on what I learned with this assignment -

This assignment was relatively easy from a coding standpoint and more difficult from a GitHub standpoint.  It has made me realize that data science is as much about learning how to interact with a community as it is just solving problems